Rename supplemental content app to resources
