+++ solution/backend/populate_content.py
@@ -8,9 +8,9 @@ def handler(event, context):
     import django
     django.setup()
 
-    call_command('loaddata', 'supplemental_content.category.json')
-    call_command('loaddata', 'supplemental_content.subcategory.json')
-    call_command('loaddata', 'supplemental_content.subpart.json')
-    call_command('loaddata', 'supplemental_content.section.json')
-    call_command('loaddata', 'supplemental_content.supplementalcontent.json')
-    call_command('loaddata', 'synonyms.json')
+    call_command('loaddata', 'resources.category.json')
+    call_command('loaddata', 'resources.subcategory.json')
+    call_command('loaddata', 'resources.subpart.json')
+    call_command('loaddata', 'resources.section.json')
+    call_command('loaddata', 'resources.supplementalcontent.json')
+    call_command('loaddata', 'search.synonyms.json')
