+++ solution/backend/regcore/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RegcoreConfig(AppConfig):
+    name = "regcore"
+    verbose_name = "Regulations core"
