+++ solution/backend/regcore/urls.py
@@ -33,7 +33,7 @@ router.register("toc", ContentsViewSet)
 urlpatterns = [
     path("v2/", include([
         path("", include('regcore.search.urls')),
-        path("", include('supplemental_content.urls')),
+        path("", include('resources.urls')),
         path("", PartsView.as_view(), name="regcore"),
         path("parser_config", ParserConfigurationView.as_view()),
         path("all_parts", PartListView.as_view()),
@@ -45,7 +45,7 @@ urlpatterns = [
         path("title/<title>/existing", ExistingPartsView.as_view()),
     ])),
     path("v3/", include([
-        path("", include('supplemental_content.v3urls')),
+        path("", include('resources.v3urls')),
         path("toc", ContentsViewSet.as_view({
             "get": "list",
         })),
