+++ solution/backend/regulations/views/reader.py
@@ -7,7 +7,7 @@ from django.urls import reverse
 from django.http import HttpResponseRedirect
 
 from regcore.models import Part
-from supplemental_content.models import Category, SubCategory
+from resources.models import Category, SubCategory
 from regulations.views.mixins import CitationContextMixin
 from regulations.views.utils import find_subpart
 from regulations.views.errors import NotInSubpart
