+++ solution/backend/resources/admin.py
