+++ solution/backend/resources/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ResourcesConfig(AppConfig):
+    name = "resources"
+    verbose_name = "Resources for regulations"
