+++ solution/backend/resources/fixtures/resources.category.json
