+++ solution/backend/resources/fixtures/resources.section.json
