+++ solution/backend/resources/mixins.py
