+++ solution/backend/resources/utils.py
