+++ solution/backend/resources/views.py
