--- solution/backend/supplemental_content/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class SupplementalContentConfig(AppConfig):
-    name = "supplemental_content"
-    verbose_name = "Supplemental content for regulations"
